# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Auth Service API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development
paths:
...
  /posts:
    get:
      summary: Get all posts
      tags: [Posts]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Create a new post
      tags: [Posts]
      security:
        - bearerAuth: [] # ← 인증 필요 추가
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
      responses:
        '201':
          description: Created — returns new post object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad Request — validation failed
        '401':
          description: Unauthorized — authentication required # ← 추가
  /posts/{id}:
    get:
      summary: Get single post by ID
      tags: [Posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique identifier of the post
      responses:
        '200':
          description: Post object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Not Found — post does not exist
    patch:
      summary: Update post by ID
      tags: [Posts]
      security:
        - bearerAuth: [] # ← 인증 필요 추가
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique identifier of the post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
      responses:
        '200':
          description: Updated — returns updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad Request — validation failed
        '401':
          description: Unauthorized — authentication required # ← 추가
        '404':
          description: Not Found — post does not exist
    delete:
      summary: Delete post by ID
      tags: [Posts]
      security:
        - bearerAuth: [] # ← 인증 필요 추가
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique identifier of the post
      responses:
        '204':
          description: No Content — post deleted successfully
        '401':
          description: Unauthorized — authentication required # ← 추가
        '404':
          description: Not Found — post does not exist
components:
  securitySchemes:
    bearerAuth: # ← 인증 스키마 추가
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatePostDto:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        authorId:
          type: string
    UpdatePostDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    Post:
      allOf:
        - $ref: '#/components/schemas/CreatePostDto'
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time